DAOImplementation acabado

+++ src/classes/Account.java
@@ -18,9 +18,9 @@ public class Account implements Serializable{
     
     private long accountId;
     private String description;
-    private float balance;
-    private float creditLine;
-    private float beginBalance;
+    private double balance;
+    private double creditLine;
+    private double beginBalance;
     private Timestamp beginBalanceTimestamp;
     private int accountType;
 
@@ -49,7 +49,7 @@ public class Account implements Serializable{
     /**
      * @param accountId the accountId to set
      */
-    public void setAccountId(int accountId) {
+    public void setAccountId(long accountId) {
         this.accountId = accountId;
     }
 
@@ -70,42 +70,42 @@ public class Account implements Serializable{
     /**
      * @return the balance
      */
-    public float getBalance() {
+    public double getBalance() {
         return balance;
     }
 
     /**
      * @param balance the balance to set
      */
-    public void setBalance(float balance) {
+    public void setBalance(double balance) {
         this.balance = balance;
     }
 
     /**
      * @return the creditLine
      */
-    public float getCreditLine() {
+    public double getCreditLine() {
         return creditLine;
     }
 
     /**
      * @param creditLine the creditLine to set
      */
-    public void setCreditLine(float creditLine) {
+    public void setCreditLine(double creditLine) {
         this.creditLine = creditLine;
     }
 
     /**
      * @return the beginBalance
      */
-    public float getBeginBalance() {
+    public double getBeginBalance() {
         return beginBalance;
     }
 
     /**
      * @param beginBalance the beginBalance to set
      */
-    public void setBeginBalance(float beginBalance) {
+    public void setBeginBalance(double beginBalance) {
         this.beginBalance = beginBalance;
     }
